usermgt/app/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
Jinja2==2.10.1
psycopg2==2.8.3
PyJWT==1.7.1
Werkzeug==0.15.5
Flask-Cors==3.0.3
